feat: provide `init` command
This command creates the simple database schema.

On Linux, the database file will be saved at `~/.local/share/books-db/books.db`. On other platforms,
it will be whatever `dirs_next::data_dir()` returns, which should be something like the `AppData`
directory on Windows.

A global `--storage-path` argument is provided to override the location, but that's really just for
easier testing.