feat: provide `add` command
The `add` command uses an ISBN number to retrieve the book details from the ISBNDb database, then
saves the book in our own database. The user is offered the chance to edit the details before the
book is saved, in which case the editor specified by the `EDITOR` or `VISUAL` environment variables
will be used.