chore: resolve clippy errors
This was the first time that clippy had been run on this codebase.

I also removed the `udeps` job from the merge workflow because I was getting a bizarre compiler
error trying to build the code when using the nightly toolchain. This is unfortunate, but I don't
feel like spending time just now trying to resolve this issue.